+++ src/main/kotlin/site/siredvin/turtlematic/computercraft/plugins/AutomataLookPlugin.kt
@@ -4,6 +4,7 @@ import dan200.computercraft.api.lua.IArguments
 import dan200.computercraft.api.lua.LuaFunction
 import dan200.computercraft.api.lua.MethodResult
 import net.minecraft.world.entity.Entity
+import net.minecraft.world.level.block.entity.BlockEntity
 import net.minecraft.world.level.block.state.BlockState
 import net.minecraft.world.phys.BlockHitResult
 import net.minecraft.world.phys.EntityHitResult
@@ -18,6 +19,7 @@ class AutomataLookPlugin(
     automataCore: BaseAutomataCorePeripheral,
     private val entityEnriches: List<KFunction2<Entity, MutableMap<String, Any>, Unit>> = emptyList(),
     private val blockStateEnriches: List<KFunction2<BlockState, MutableMap<String, Any>, Unit>> = emptyList(),
+    private val blockEntityEnriches: List<KFunction2<BlockEntity, MutableMap<String, Any>, Unit>> = emptyList(),
     private val allowedMods: Set<InteractionMode> = InteractionMode.values().toSet()
 ) : AutomataCorePlugin(automataCore) {
 
@@ -31,6 +33,10 @@ class AutomataLookPlugin(
         val base = LuaRepresentation.forBlockState(blockState)
         blockStateEnriches.forEach { it.invoke(blockState, base) }
         return base
+    }
+
+    private fun blockEntityConverter(blockEntity: BlockEntity, data: MutableMap<String, Any>) {
+        blockEntityEnriches.forEach { it.invoke(blockEntity, data) }
     }
 
     private fun lookImpl(arguments: IArguments): MethodResult {
@@ -48,8 +54,13 @@ class AutomataLookPlugin(
         }, overwrittenDirection=overwrittenDirection?.minecraftDirection)
         if (result.type == HitResult.Type.MISS)
             return MethodResult.of(null, "Nothing found")
-        if (result is BlockHitResult)
-            return MethodResult.of(blockStateConverter(owner.level!!.getBlockState(result.blockPos)))
+        if (result is BlockHitResult) {
+            val base = blockStateConverter(owner.level!!.getBlockState(result.blockPos))
+            val entity = owner.level!!.getBlockEntity(result.blockPos)
+            if (entity != null)
+                blockEntityConverter(entity, base)
+            return MethodResult.of(base)
+        }
         if (result is EntityHitResult)
             return MethodResult.of(entityConverter(result.entity))
         return MethodResult.of(null, "Nothing found")